src/importer/wikipedia/requirements.txt
datasets
apache_beam
mwparserfromhell
wikitextparser
farm-haystack